# api/openapi.yaml
openapi: 3.1.0
info:
  title: Go-Stock Local Web API
  version: 1.7.3
  description: Local-only Web API for market data, watchlists, funds, settings, and AI research.
servers:
  - url: http://127.0.0.1:34115
paths:
...
  /api/v1/system/update-check:
    post:
      operationId: checkForUpdates
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdateCheckRequest"}}}}
      responses: {"202": {description: Update check accepted, content: {application/json: {schema: {$ref: "#/components/schemas/AcceptedResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
  /api/v1/system/shutdown:
    post:
      operationId: shutdownSystem
      responses: {"200": {description: Shutdown requested, content: {application/json: {schema: {$ref: "#/components/schemas/CommandResponse"}}}}, "403": {$ref: "#/components/responses/Forbidden"}}
  /api/v1/settings:
    get: {operationId: getSettings, responses: {"200": {description: Settings}}}
    put:
      operationId: updateSettings
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/JsonObject"}}}}
      responses: {"200": {description: Settings updated, content: {application/json: {schema: {$ref: "#/components/schemas/CommandResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
...
  /api/v1/ai/configs/test:
    post:
      operationId: testAIConfig
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/AIConfigTestRequest"}}}}
      responses: {"200": {description: AI configuration test result}, "400": {$ref: "#/components/responses/BadRequest"}}
  /api/v1/groups:
    get: {operationId: listGroups, responses: {"200": {description: Groups}}}
    post:
      operationId: createGroup
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/JsonObject"}}}}
      responses: {"200": {description: Group created, content: {application/json: {schema: {$ref: "#/components/schemas/CommandResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
...
  /api/v1/groups/{id}/sort:
    put:
      operationId: updateGroupSort
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/GroupSortRequest"}}}}
      responses: {"200": {description: Group sort updated, content: {application/json: {schema: {$ref: "#/components/schemas/CommandResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
  /api/v1/groups/{id}/stocks:
    post:
      operationId: addGroupStock
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/StockCodeRequest"}}}}
      responses: {"200": {description: Stock added to group, content: {application/json: {schema: {$ref: "#/components/schemas/CommandResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
...
  /api/v1/research/analysis-runs:
    get:
      operationId: listAnalysisRuns
      responses:
        "200": {description: Analysis runs, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/AnalysisRunSummary"}}}}}
    post:
      operationId: startAnalysisRun
      responses: {"202": {description: Analysis accepted, content: {application/json: {schema: {$ref: "#/components/schemas/AcceptedResponse"}}}}, "409": {description: Analysis already running, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}}
...
  /api/v1/research/recommendations/{id}/chart/refresh:
    post:
      operationId: refreshRecommendationChart
      description: Fills missing minute coverage using only enabled providers, refreshes the quote, and returns the chart data.
      responses:
        "200": {description: Refreshed professional chart data, content: {application/json: {schema: {$ref: "#/components/schemas/RecommendationChart"}}}}
        "409": {description: The same recommendation is already being refreshed, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
  /api/v1/exports/config:
    post:
      operationId: exportConfig
      requestBody: {content: {application/json: {schema: {$ref: "#/components/schemas/ExportRequest"}}}}
      responses: {"200": {description: Export, content: {application/json: {schema: {$ref: "#/components/schemas/ExportResponse"}}}}, "400": {$ref: "#/components/responses/BadRequest"}}
...
components:
  responses:
    BadRequest:
      description: Invalid JSON request or parameters
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    Forbidden:
      description: Request is not from the bound loopback origin
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
  schemas:
    JsonObject:
      type: object
      description: Provider- or configuration-defined JSON object whose stable fields are not owned by this transport layer.
    ErrorResponse:
      type: object
      required: [error]
      properties: {error: {type: string}}
    CommandResponse:
      type: object
      required: [ok]
      properties: {ok: {type: boolean}, message: {type: string}}
    AcceptedResponse:
      type: object
      required: [accepted]
      properties: {accepted: {type: boolean}}
    UpdateCheckRequest:
      type: object
      required: [flag]
      properties: {flag: {type: integer}}
    AIConfigTestRequest:
      type: object
      required: [id]
      properties: {id: {type: integer}}
    GroupSortRequest:
      type: object
      required: [sort]
      properties: {sort: {type: integer}}
    StockCodeRequest:
      type: object
      required: [stockCode]
      properties: {stockCode: {type: string}}
...
    ExportRequest:
      type: object
      properties: {mode: {type: string, enum: [download, server]}}
    ExportResponse:
      type: object
      required: [mode, filename]
      properties:
        mode: {type: string, enum: [download, server]}
        filename: {type: string}
        mime: {type: string}
        contentBase64: {type: string}
        path: {type: string}
...
    AnalysisRunSummary:
      type: object
      required: [runId, scheduledFor, startedAt, status, recommendationCount, sourceCount, failedSourceCount]
      properties:
        runId: {type: string}
        scheduledFor: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time}
        status: {type: string}
        providerName: {type: string}
        modelName: {type: string}
        recommendationCount: {type: integer}
        failureReason: {type: string}
        sourceCount: {type: integer}
        failedSourceCount: {type: integer}
...
    RecommendationChartBar:
      type: object
      required: [at, open, high, low, close, volume, amount, source, netPnl, netYieldRate]
      properties:
        at: {type: string, format: date-time}
        open: {type: number}
        high: {type: number}
        low: {type: number}
        close: {type: number}
        volume: {type: number}
        amount: {type: number}
        source: {type: string}
        netPnl: {type: number, description: Full-cost estimated net profit at this minute.}
        netYieldRate: {type: number, description: Full-cost estimated net yield at this minute.}
    RecommendationChartTrade:
      type: object
      required: [side, tradedAt, marketPrice, executionPrice, quantity, totalFees, netCashFlow, markerSnapped]
      properties:
        side: {type: string, enum: [buy, sell]}
        tradedAt: {type: string, format: date-time}
        marketPrice: {type: number}
        executionPrice: {type: number}
        quantity: {type: integer, format: int64}
        totalFees: {type: number}
        netCashFlow: {type: number}
        markerAt: {type: string, format: date-time, description: Nearest bar in the same trading segment within five minutes; omitted if unavailable.}
        markerSnapped: {type: boolean}
    RecommendationChartSession:
      type: object
      required: [date, previousClose, status]
      properties:
        date: {type: string, format: date}
        previousClose: {type: number}
        status: {type: string, enum: [complete, partial, missing]}
    ChartProviderError:
      type: object
      required: [provider, message]
      properties:
        provider: {type: string}
        message: {type: string}
    RecommendationChart:
      type: object
      required: [recommendationId, stockCode, stockName, status, rangeFrom, rangeTo, refreshedAt, currentPrice, currentNetPnl, currentNetYieldRate, missingSessions, providerErrors, sessions, bars, trades]
      properties:
        recommendationId: {type: string}
        stockCode: {type: string}
        stockName: {type: string}
        status: {type: string, enum: [complete, partial, empty]}
        rangeFrom: {type: string, format: date-time}
        rangeTo: {type: string, format: date-time}
        refreshedAt: {type: string, format: date-time}
        quoteAt: {type: string, format: date-time}
        currentPrice: {type: number}
        currentNetPnl: {type: number}
        currentNetYieldRate: {type: number}
        missingSessions: {type: array, items: {type: string, format: date}}
        providerErrors: {type: array, items: {$ref: "#/components/schemas/ChartProviderError"}}
        sessions: {type: array, items: {$ref: "#/components/schemas/RecommendationChartSession"}}
        bars: {type: array, items: {$ref: "#/components/schemas/RecommendationChartBar"}}
        trades: {type: array, items: {$ref: "#/components/schemas/RecommendationChartTrade"}}
...